+++ src/config.ts
@@ -4,6 +4,7 @@ export interface ExposedUserConfig {
   include?: string[];
   outDir?: string;
   initializers?: boolean;
+  imports?: boolean;
 }
 
 export interface ExposedConfig extends ExposedUserConfig {