+++ src/core.test.ts
@@ -12,13 +12,14 @@ const baseConfig = {
 };
 
 test('snapshot', async t => {
-  const exposedPath = getExposedPath(hre.config);
-  const rootRelativeExposedPath = path.relative(hre.config.paths.root, exposedPath);
+  const rootRelativeSourcesPath = path.relative(hre.config.paths.root, hre.config.paths.sources);
+  const sourcesPathPrefix = path.normalize(rootRelativeSourcesPath + '/');
+  const include = (sourceName: string) => sourceName.startsWith(sourcesPathPrefix);
 
   const [bip] = await hre.artifacts.getBuildInfoPaths();
   const bi: BuildInfo = JSON.parse(await fs.readFile(bip!, 'utf8'));
   const config = { paths: hre.config.paths, exposed: { ...baseConfig, initializers: false } };
-  const exposed = getExposed(bi.output, sourceName => !sourceName.startsWith(rootRelativeExposedPath), config);
+  const exposed = getExposed(bi.output, include, config);
   const exposedFiles = [...exposed.values()].sort((a, b) => a.absolutePath.localeCompare(b.absolutePath))
   for (const rf of exposedFiles) { 
     t.snapshot(rf.content.rawContent);
@@ -33,5 +34,20 @@ test('snapshot initializers', async t => {
   const exposedFiles = [...exposed.values()].sort((a, b) => a.absolutePath.localeCompare(b.absolutePath))
   for (const rf of exposedFiles) {
     t.snapshot(rf.content.rawContent);
+  }
+});
+
+test('snapshot imports', async t => {
+  const rootRelativeSourcesPath = path.relative(hre.config.paths.root, hre.config.paths.sources);
+  const sourcesPathPrefix = path.normalize(rootRelativeSourcesPath + '/');
+
+  const [bip] = await hre.artifacts.getBuildInfoPaths();
+  const bi: BuildInfo = JSON.parse(await fs.readFile(bip!, 'utf8'));
+  const config = { paths: hre.config.paths, exposed: { ...baseConfig, initializers: false, imports: true } };
+  const exposed = getExposed(bi.output, sourceName => sourceName === 'contracts/Imported.sol', config);
+  const exposedFiles = [...exposed.values()].sort((a, b) => a.absolutePath.localeCompare(b.absolutePath))
+  for (const rf of exposedFiles) { 
+    const { absolutePath, content: { rawContent } } = rf;
+    t.snapshot({ absolutePath, rawContent });
   }
 });